requirements.txt
# LiveKit Agents and plugins
livekit-agents[deepgram,openai,cartesia,silero,turn-detector,ai4bharat,elevenlabs]==1.2.5
livekit-plugins-noise-cancellation==0.2.5
livekit-api==1.0.5

# Core dependencies
python-dotenv==1.0.0
aiohttp==3.12.15
